addons/images/map_query/requirements.txt
DateTime==5.1
pymongo==4.3.3
python-dateutil==2.8.2
pytest-cov==4.0.0
kafka-python==2.0.2